requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
astor==0.8.1
audioread==2.1.8
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
efficientnet==1.1.0
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.6
imageio==2.9.0
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.1.1
matplotlib==3.3.0
networkx==2.4
numba==0.48.0
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==1.0.5
Pillow==7.2.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pydot==1.4.1
pydub==0.24.1
pygobject==3.26.1
pyparsing==2.4.7
python-apt==1.6.4
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.25
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
SecretStorage==2.3.1
six==1.13.0
sklearn==0.0
SoundFile==0.10.3.post1
tensorboard==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.7.24
tqdm==4.47.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
yaspin==1.0.0
